gitconfig

+++ run.sh
@@ -63,11 +63,11 @@ _git_config() {
   GIT_USEREMAIL="${ANSWER:-${DEFAULT}}"
   git config --global user.email "${GIT_USEREMAIL}"
 
-  git config --global core.autocrlf input
-  git config --global core.pager ''
-  git config --global core.precomposeunicode true
-  git config --global core.quotepath false
-  git config --global pull.ff only
+  # git config --global core.autocrlf input
+  # git config --global core.pager ''
+  # git config --global core.precomposeunicode true
+  # git config --global core.quotepath false
+  # git config --global pull.ff only
 
   _command "git config --list"
   git config --list
@@ -178,13 +178,14 @@ fi
 _backup ~/.aliases
 curl -fsSL -o ~/.aliases https://raw.githubusercontent.com/nalbam/dotfiles/main/.aliases
 
-# .gitconfig
-_backup ~/.gitconfig
-curl -fsSL -o ~/.gitconfig https://raw.githubusercontent.com/nalbam/dotfiles/main/.gitconfig
-
 # .vimrc
 _backup ~/.vimrc
 curl -fsSL -o ~/.vimrc https://raw.githubusercontent.com/nalbam/dotfiles/main/.vimrc
+
+# .gitconfig
+if [ ! -f ~/.gitconfig ]; then
+  curl -fsSL -o ~/.gitconfig https://raw.githubusercontent.com/nalbam/dotfiles/main/.gitconfig
+fi
 
 # git config
 GIT_USERNAME="$(git config --global user.name)"
